:recycle: Refactor GameResult
게임 결과, 게임 룸 리팩토링중...
